% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmodel.R
\name{bestmodel}
\alias{bestmodel}
\title{Choose the best model}
\usage{
bestmodel(quickmodel)
}
\arguments{
\item{quickmodel}{A \code{quickmodel} object}
}
\value{
The model with highest accuracy.
}
\description{
\code{bestmodels} will find the best models from a \code{quickmodel} object.
}
\details{
This function will take the result from quickmodel and pick the best model.
It will pick by the lowest RMSE for regression and highest Accuracy for classification.
Notes: This function now only considers RMSE and Accuracy as the metric.
}
\examples{
# classification models
data("PIMA", package="regclass")
x=quickmodel(Diabetes~., data = PIMA)
bm=bestmodel(x)
# regression models
data(Boston, package="MASS")
x=quickmodel(medv~., data = Boston)
bm=bestmodel(x)
}
