% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmodel.R
\name{quickmodel}
\alias{quickmodel}
\title{Train multiple machine learning models}
\usage{
quickmodel(
  formula,
  data,
  metric,
  methods,
  trControl = trainControl(),
  tuneGrid = NULL,
  tuneLength = 3,
  partition = 0.8,
  seed = 1234,
  ...
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{Data frame from which variables specified in formula are preferentially to be taken.}

\item{metric}{A string that specifies what summary metric will be used to select the optimal model. By default, possible values are "RMSE" and "Rsquared" for regression and "Accuracy" for classification.}

\item{methods}{A character vector specifying which classification or regression model to use. By default, possible values are \code{c("lm", "knn", "rf", "rpart", "gbm", "glmnet")} for regression and \code{c("glm", "knn", "rf", "rpart", "gbm")} for classification.}

\item{trControl}{A list of values that define how this function acts. Default: trainControl()}

\item{tuneGrid}{A data frame with possible tuning values. The columns are named the same as the tuning parameters in each method preceded by a period (e.g. .decay, .lambda)., Default: NULL}

\item{tuneLength}{an integer denoting the number of levels for each tuning parameter that should be generated by createGrid. (NOTE: If given, this argument must be named.), Default: 3}

\item{partition}{The partition rate. Default: 0.8}

\item{seed}{The seed to do a random partition. Default: 1234}

\item{...}{Optional parameters for \code{train} in \code{caret}}
}
\value{
A list of class \code{quickmodel} with 5 components:
\describe{
\item{train}{The train dataset}
\item{test}{The test dataset}
\item{models}{A list of models}
\item{methods}{A character vector of machine learning models}
\item{metric}{The selected metric. Default: "RMSE" for regression and "Accuracy" for classification.}
}
}
\description{
\code{quickmodel} will build multiple machine learning models with one function call.
}
\examples{
# classification model
data(PIMA)
class <- quickmodel(Diabetes ~ ., data = PIMA)

# regression model
data(Boston)
reg <- quickmodel(medv ~ ., data = Boston)
}
