% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmodel.R
\name{bestmodel}
\alias{bestmodel}
\title{Choose the best model}
\usage{
bestmodel(quickmodel)
}
\arguments{
\item{quickmodel}{A \code{quickmodel} object.}
}
\value{
The best model based on the evaluation metric.
}
\description{
\code{bestmodel} will retrieve the best model from a \code{quickmodel} object.
}
\details{
This function will take the results from quickmodel and pick the best model.
It will pick by the lowest RMSE for regression and highest Accuracy for classification.
}
\examples{
# classification model
data(PIMA)
class <- quickmodel(Diabetes ~ ., data = PIMA)
bm <- bestmodel(class)

# regression model
data(Boston)
reg <- quickmodel(medv ~ ., data = Boston)
bm <- bestmodel(reg)
}
